Create core programmatically and download core config dovecot-github
This commit adds a file to the root directory of the image to be run
on image creation.

It creates the core and copies the dovecot-recommended schema to the
new cores' data directory. Other than the schema, no "hand-written"
configs are included.

This is still WIP, since the file ought to be removed after creation
and I am unsure if the file should be the new ENTRYPOINT of the
container (with modifications).