Extract last IP address by default
This takes care of the case where multiple IP addresses exist for a container and ensures that get_container_ip only returns one (the last one).

The issue with the existing file occurs when multiple IP addresses exist for a service.  For example, if nginx-mailcow is also on a network with a reverse proxy server.  In this case, the get_container_ip function would return a string with a space separated list of IP addresses.  Check functions calling get_container_ip would be unable to reach their respective services with this non-ip address string and would show these services as failed.