Watchdog: Allow sending notifications via Slack webhooks
With this commit we allow watchdog notifications to be sent to a Slack
channel via a Slack webhook (https://api.slack.com/messaging/webhooks).

Notifications are sent to all defined methods. That if both watchdog
e-mail and slack webhook is defined, notifications are sent to both. If
only one of them are defined, only one notification method will be used.

In order to do so, a bunch of if statements were moved to the
`mail_error` function.