Correct syntax for MX and SRV Records
Per spec, both of those records need to have a priority / preference[^1]. And, in the
case of the SRV-Record[^2], it also needs to provide a weight and the
Port **infront** of the domain name. The Validating of the DNS-Entries
is also adjusted to pass as long as the target matches without the need
for a matching priority.

As an example for the domain domain.tld the current behaviour is:

```zone
@	IN	MX	mailsrv.domain.tld.
_autodiscover._tcp	IN	SRV	mailsrv.domain.tld. 443
```

This is bad syntax, the correct on, and as this patch introduces it is
like this:

```zone
@	IN	MX	10 mailsrv.domain.tld.
_autodiscover._tcp	IN	SRV	10 10 443 mailsrv.domain.tld.
```

[^1]: https://tools.ietf.org/html/rfc5321#page-70
[^2]: https://tools.ietf.org/html/rfc2782