Fixed restore with NODATE
Apparently I changed the problematic for loop on the VM, but I forgot to copy the script back to the git folder before committing. I put the for loop into the IF branches now.

TESTING IS UNDERWAY - I will comment again once it is 100% verified working (for me, again, sorry.)