Add coreutils version check to ensure compatibility with all distros
The -n flag deprecation is only present in coreutils >= 9.3, and the replacement flag is only present in versions older than this. A conditional is added to ensure that the new version of the command only runs on systems with coreutils of an appropriate version, using the original on older systems.